Simplify `commonerrors` package
